refactor:  address feedback
