+++ src/features/user/MFV2/TpoProjects/index.tsx
@@ -77,7 +77,7 @@ export default function ProjectsContainer({ profile }: Props) {
                   >
                     <ProjectSnippet
                       project={project.properties}
-                      displayPopup={true}
+                      shouldDisplayPopup={false}
                     />
                   </div>
                 );
